% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.r
\docType{class}
\name{ClusterInput-class}
\alias{ClusterInput-class}
\title{Class \code{ClusterInput}}
\description{
This \code{ClusterInput} class holds the inferred clonotype lineages from \code{getClonotypeLineages}, which can be used as input for the following clustering analysis.
}
\section{Slots}{

\describe{
\item{\code{lineages.}}{A list of lineages corresponding the the clonotypes}
}}

\examples{
# TrajectoryDataSet objects are generated from \code{getClonotypeLineages}
}
