% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.r, R/HelperFuntions.r
\docType{class}
\name{CombinedDataSet-class}
\alias{CombinedDataSet-class}
\alias{reducedDim,CombinedDataSet-method}
\title{Class \code{CombinedDataSet}}
\usage{
\S4method{reducedDim}{CombinedDataSet}(CombinedDataSet)
}
\arguments{
\item{combined}{a \code{CombinedDataSet} object}
}
\description{
This \code{CombinedDataSet} class holds the combined scRNA-seq and scTCR-seq information for cells, which are relevant for 
	performing clonotype lineage inference and visualization.
}
\section{Methods (by generic)}{
\itemize{
\item \code{reducedDim}: returns the matrix representing the reduced
dimensional dataset,i.e., UMAP coordinates
}}

\section{Slots}{

\describe{
\item{\code{seurat.}}{A seurat object containing the UMAP coordinates, cluster labels and clonotype information for cells
\itemize{
\item{cdr3} {character. Vector of CDR3 sequence that are used to define cell clonotype }
\item{cell.id}{character. Vector of cell barcodes}
\item{clusters}{character or numeric. A vector of character or numeric values indicating clustering labels for cells}
\item{Other terms} {other metadata stored in the Seurat object}}}
}}

