% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.r, R/getClonotypeLineages.R
\name{getClonotypeLineages}
\alias{getClonotypeLineages}
\alias{getClonotypeLineages,CombinedDataSet-method}
\title{Infer clonotype lineage structure from \code{CombinedDatSet}}
\usage{
getClonotypeLineages(CombinedDataSet, ...)

\S4method{getClonotypeLineages}{CombinedDataSet}(
  CombinedDataSet,
  start.clus = NULL,
  end.clus = NULL,
  dist.method = "simple",
  use.median = TRUE,
  ...
)
}
\arguments{
\item{CombinedDataSet}{The \code{CombinedDataSet} object obtained from \code{getCombinedDataSet}}

\item{...}{Additional arguments to specify how lineages are constructed from
clusters.}

\item{start.clus}{(optional) character, indicates the starting cluster(s)
from which lineages will be drawn.}

\item{end.clus}{(optional) character, indicates which cluster(s) will be
forced to be leaf nodes in the graph.}

\item{dist.method}{(optional) character, specifies the method for calculating
distances between clusters. Default is \code{"simple"}, see
\code{\link[TrajectoryUtils]{createClusterMST}} for details.}

\item{use.median}{logical, whether to use the median (instead of mean) when
calculating cluster centroid coordinates. Default is \code{'TRUE'}.}

\item{data}{A class of \code{CombinedDataSet} object}
}
\value{
A class of \code{TrajectoryDataSet} object
}
\description{
This function infers the clonotype lineages structure from \code{CombinedDataSet}.
}
\details{
Given a \code{CombinedDataSet} object, this function constructs the minimum spanning tree(s) on cells sharing the same clonotype.
 These cells may belong to different clusters defined by scRNA-seq features, therefore a graph can be constructed with node being the center of a cluster,
 and weight of an edge being the distance between the nodes. MST is built based on such a graph.

Once the MST is known, lineages are identified in
  any tree with at least two clusters. For a given tree, if there is an
  annotated starting cluster, every possible path out of a starting cluster
  and ending in a leaf that isn't another starting cluster will be returned.
  If no starting cluster is annotated, one will be chosen by a heuristic
  method, but this is not recommended.
}
\examples{
TCR <-read.csv('/PATH/TO/YOUR/scTCR-seqData/',header=T)
load('Mice.sub.rda')
Combined <- getCombinedDataSet(TCR,Mice.sub)
Trajectory <- getClonotypeLineages(Combined,start.clus = NULL, end.clus = NULL, dist.method = 'simple', use.median = TRUE)

}
