% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusteringInput.r
\name{getClusteringInput,TrajectoryDataSet-method}
\alias{getClusteringInput,TrajectoryDataSet-method}
\title{get clustering input}
\usage{
\S4method{getClusteringInput}{TrajectoryDataSet}(TrajectoryDataSet)
}
\arguments{
\item{object}{A \code{TrajectoryDataSet} object, which contains a list of clonotype lineages}
}
\value{
A object of class \code{ClusterInput} with clonotype information as well as UMAP coordinates.
}
\description{
This function prepare the input for the \code{clustLineages} function
}
\examples{
TCR <-read.csv("/PATH/TO/YOUR/scTCR-seqData/",header=T)
load('Mice.sub.rda')
Combined <- getCombinedDataSet(TCR,Mice.sub)
Trajectory <- getClonotypeLineages(Combined,start.clus = NULL, end.clus = NULL, dist.method = 'simple', use.median = TRUE)
ClusterInput <-getClusteringInput(Trajectory)

}
