% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.r
\name{getClusteringInput}
\alias{getClusteringInput}
\title{prepare the input for \code{clustLineages}}
\usage{
getClusteringInput(TrajectoryDataSet, ...)
}
\arguments{
\item{TrajectoryDataSet}{The \code{TrajectoryDataSet} object obtained from \code{getClonotypeLineages}}
}
\description{
prepare the input for \code{clustLineages}
}
