% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShinyInput.r
\name{getShinyInput,CombinedDataSet-method}
\alias{getShinyInput,CombinedDataSet-method}
\title{prepare Shiny input}
\usage{
\S4method{getShinyInput}{CombinedDataSet}(CombinedDataSet)
}
\arguments{
\item{object}{A \code{CombinedDataSet} object}
}
\value{
A data frame with cell clonotype and UMAP coordinates .
}
\description{
This function prepares the input for the two shiny apps
}
\details{
This function takes the \code{CombinedDataSet}object as the input and extract the UMAP coordinates for each clonotype, which will be used as input of the two shiny apps.
}
\examples{
TCR <-read.csv("/PATH/TO/YOUR/scTCR-seqData/",header=T)
load('Mice.sub.rda')
Combined <- getCombinedDataSet(TCR,Mice.sub)
Trajectory <- getClonotyoeLineages(Combined,start.clus = NULL, end.clus = NULL, dist.method = 'simple', use.median = TRUE)
clustInput <- getClusteringInput(Trajectory)
shinyInput <- shinyInput(clustInput)
write.csv(shinyInput,file='/path/to/folder/shiny.csv')


}
