% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{list.input.return}
\alias{list.input.return}
\title{generate a list from CombinedDataSet object}
\usage{
list.input.return(Combined, group.by)
}
\arguments{
\item{Combined}{a \code{CombinedDataSet}object.}

\item{group.by}{The column header to group the new list. NULL will return clusters.}
}
\description{
generate a list from CombinedDataSet object
}
