% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyClust.r
\name{shinyClust}
\alias{shinyClust}
\title{cluster inferred clonotype trajectories}
\usage{
shinyClust(TrajectoryDataSet = NULL, CombinedDataSet = NULL)
}
\arguments{
\item{Combined}{a \code{CombinedDataSet} object}

\item{trajectory}{a \code{TrajectoryDataSet} object}
}
\description{
cluster inferred clonotype trajectories
}
\examples{
TCR <-read.csv("/PATH/TO/YOUR/scTCR-seqData/",header=T)
load('Mice.sub.rda')
Combined <- getCombinedDataSet(TCR,Mice.sub)
Trajectory <- getClonotypeLineages(Combined,start.clus = NULL, end.clus = NULL, dist.method = 'simple', use.median = TRUE)
shinyClust(Trajectory,Combined)
}
